// frontend-legacy/src/app/compliance/shared/type/compliance-standard-section.type.ts
import {ComplianceStandardType} from './compliance-standard.type';

export class ComplianceStandardSectionType {
  id: number;
  standard: ComplianceStandardType;
  standardId: number;
  standardSectionDescription: string;
  standardSectionName: string;
  isCollapsed?: boolean;
  isActive?: boolean;
}